Rename component to better match its purpose
